zaakafhandelcomponent-1165 Taak aanvullende informatie uitbreiden met mogelijkheid om termijn in ...

+++ src/main/java/net/atos/zac/app/zaken/model/RESTZaakOpschortGegevens.java
@@ -13,7 +13,7 @@ public class RESTZaakOpschortGegevens {
 
     public String redenOpschorting;
 
-    public Integer duurDagen;
+    public long duurDagen;
 
     public LocalDate einddatumGepland;
